Fix clippy warning and restore process_display_mode_change signature
- Drop the unnecessary `as i32` cast in server/mod.rs (haversine helper
  already returns i32) — clippy was failing with -D warnings
- Revert the ctx parameter on process_display_mode_change to keep its
  existing test call sites compiling; trigger the WASM-only statistics
  refresh from the SetDisplayMode event handler instead (which already
  has ctx in scope)

+++ src/server/mod.rs
@@ -182,7 +182,7 @@ async fn get_history(State(state): State<SharedState>) -> impl IntoResponse {
                     state.airport_by_icao.get(&h.arrival_icao),
                 ) {
                     (Some(dep), Some(arr)) => {
-                        crate::util::calculate_haversine_distance_nm(dep, arr) as i32
+                        crate::util::calculate_haversine_distance_nm(dep, arr)
                     }
                     _ => 0,
                 }
